# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: CRCI Node API
  version: "0.1.0"
  description: >
    REST and WebSocket API for a CRCI mesh node.
    Exposes node health, Merkle chain state, and divergence alerts.
servers:
  - url: http://localhost:8080
    description: Local node
paths:
...
  /divergences/{peer_id}:
    get:
      summary: List divergence alerts for a specific peer
      operationId: listDivergencesByPeer
      parameters:
        - name: peer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of divergence alerts for the peer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DivergenceAlert"
        "404":
          description: Peer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /simulate/panic:
    post:
      summary: Simulate a panic event
      operationId: simulatePanic
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Panic simulation queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:
                    type: boolean
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported media type (must be application/json)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    DivergenceAlert:
      type: object
      required:
        - id
        - detected_at_ms
        - peer_id
        - divergence_seq
        - expected_hash
        - received_hash
      properties:
        id:
          type: string
        detected_at_ms:
          type: integer
        peer_id:
          type: string
        divergence_seq:
          type: integer
        expected_hash:
          type: string
        received_hash:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
        code:
          type: string